- Make getObject() use the same datatypes as connector/j does (as documented here : http://dev.mysql.com/doc/refman/5.0/en/connector-j-reference-type-conversions.html) - Introduce connection parameters analog to Connector/J to fine-tune datatype mapping (tiny1isBit and yearIsDateType) - Add an extensive unit test to check consistency between getObject(), ResultSetMetadata.getColumnClassName() and ResultSetMatadata.getTypeName() - Fix broken cases in the test suite. - Fix parsing of stored procedure metadata (lookup for IN/OUT/INOUT should be case-insensitive)

+++ src/test/java/org/skysql/jdbc/ResultSetMetaDataTest.java
@@ -31,8 +31,8 @@ public class ResultSetMetaDataTest extends BaseTest {
         assertEquals(ResultSetMetaData.columnNullable,rsmd.isNullable(2));
         assertEquals(ResultSetMetaData.columnNoNulls,rsmd.isNullable(1));
         assertEquals(String.class.getName(), rsmd.getColumnClassName(2));
-        assertEquals(Long.class.getName(), rsmd.getColumnClassName(1));
-        assertEquals(Long.class.getName(), rsmd.getColumnClassName(3));
+        assertEquals(Integer.class.getName(), rsmd.getColumnClassName(1));
+        assertEquals(Integer.class.getName(), rsmd.getColumnClassName(3));
         assertEquals("id_col",rsmd.getColumnLabel(1));
         assertEquals("nullable_col",rsmd.getColumnLabel(2));
         assertEquals("something",rsmd.getColumnLabel(3));
